- Fix the build on 10.8 - Begin stubbing out a TRString formatting replacement

+++ trunk/src/TRString.m
@@ -206,7 +206,7 @@
         return (false);
     }
 
-    *value = i;
+    *value = (int) i;
     return (true);
 }
